feat: separate ToolsContainer and ToolContainer
