+++ plugins/toolbox/src/components/Root/Root.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import { ToolsPage, ToolsPageProps } from './ToolsPage';
+import { ToolsPage } from './ToolsPage';
 import { ToolPage } from './ToolPage';
+import type { ToolsContainerProps } from './ToolsContainer';
 
-export const Root = (props: ToolsPageProps) => {
+export const Root = (props: ToolsContainerProps) => {
   return (
     <Routes>
       <Route path="/" element={<ToolsPage {...props} />} />
