+++ plugins/toolbox/src/components/Root/ToolPage.tsx
@@ -1,26 +1,11 @@
-import React, { Suspense } from 'react';
-import { ToolsPageProps } from './ToolsPage';
+import React from 'react';
+import { ToolsContainerProps } from './ToolsContainer';
 import { defaultTools } from './tools';
 import { useParams } from 'react-router-dom';
-import { ContentHeader } from '@backstage/core-components';
-import Box from '@mui/material/Box';
-import CircularProgress from '@mui/material/CircularProgress';
 import { useToolboxTranslation } from '../../hooks';
-import { styled } from '@mui/material/styles';
+import { ToolContainer } from './ToolContainer';
 
-const ToolContainer = styled('div')(({ theme }) => ({
-  padding: '1rem',
-  position: 'absolute',
-  top: '0px',
-  left: '0px',
-  width: '100%',
-  height: '100%',
-  overflow: 'auto',
-  zIndex: 10000,
-  backgroundColor: theme.palette.background.default,
-}));
-
-export const ToolPage = (props: ToolsPageProps) => {
+export const ToolPage = (props: ToolsContainerProps) => {
   const { extraTools } = props;
   const params = useParams();
   const { t } = useToolboxTranslation();
@@ -44,25 +29,12 @@ export const ToolPage = (props: ToolsPageProps) => {
   });
 
   return (
-    <ToolContainer>
-      <Suspense
-        fallback={
-          <Box
-            display="flex"
-            width="100%"
-            height="50%"
-            alignItems="center"
-            justifyContent="center"
-          >
-            <CircularProgress />
-          </Box>
-        }
-      >
-        <ContentHeader title={title} description={description}>
-          {tool.headerButtons}
-        </ContentHeader>
-        {tool.component}
-      </Suspense>
-    </ToolContainer>
+    <ToolContainer
+      tool={{
+        ...tool,
+        name: title,
+        description,
+      }}
+    />
   );
 };
