+++ plugins/toolbox/src/index.ts
@@ -1,4 +1,10 @@
-export { toolboxPlugin, ToolboxPage, ToolboxHomepageCard } from './plugin';
+export {
+  toolboxPlugin,
+  ToolboxPage,
+  ToolsContainer,
+  ToolContainer,
+  ToolboxHomepageCard,
+} from './plugin';
 export * from '@drodil/backstage-plugin-toolbox-react';
 export * from './components/Root';
 export * from './components/DefaultEditor';
